changes for version 2.0

+++ src/main/java/cz/cvut/fel/aic/apdemo/DriveAgent.java
@@ -84,7 +84,7 @@ public class DriveAgent extends Agent implements Driver<PhysicalVehicle> {
 	}
 
 	@Override
-	public double getVelocity() {
+	public int getVelocity() {
 		return 15;
 	}
